Misc fixes
Fix anti-pop function

Fix dialer exiting when voice command ends and a call is immediately started.
This occurs when the phones voice command is asked to dial a call. It hangs up it's
session and starts the call.  Our delayed hangup was ending and missing the start
of the second call.